function_app/requirements.txt
# Azure Functions Core
azure-functions>=1.18.0,<2.0.0

# Azure SDK for Storage
azure-identity>=1.15.0,<2.0.0
azure-storage-blob>=12.19.0,<13.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0
